chore: move genesis file
